Move sandbox functions to dedicated Temporal queue
