+++ front/temporal/agent_loop/client.ts
@@ -2,9 +2,6 @@ import type { Authenticator } from "@app/lib/auth";
 import { DustError } from "@app/lib/error";
 import type { AgentMCPActionResource } from "@app/lib/resources/agent_mcp_action_resource";
 import { ConversationResource } from "@app/lib/resources/conversation_resource";
-import type { SandboxFunctionInvocationResource } from "@app/lib/resources/sandbox_function_invocation_resource";
-import type { SandboxFunctionMCPActionResource } from "@app/lib/resources/sandbox_function_mcp_action_resource";
-import type { SandboxFunctionResource } from "@app/lib/resources/sandbox_function_resource";
 import { getTemporalClientForAgentNamespace } from "@app/lib/temporal";
 import logger from "@app/logger/logger";
 import { logAgentLoopStart } from "@app/temporal/agent_loop/activities/instrumentation";
@@ -12,8 +9,6 @@ import {
   makeAgentLoopWorkflowId,
   makeCompactionWorkflowId,
   makeSandboxChildToolWorkflowId,
-  makeSandboxFunctionInvocationWorkflowId,
-  makeSandboxFunctionToolWorkflowId,
 } from "@app/temporal/agent_loop/lib/workflow_ids";
 import type {
   AgentLoopArgs,
@@ -23,15 +18,12 @@ import type { CompactionSourceConversation } from "@app/types/assistant/compacti
 import type { SupportedModel } from "@app/types/assistant/models/types";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
-import { normalizeError } from "@app/types/shared/utils/error_utils";
 import { WorkflowExecutionAlreadyStartedError } from "@temporalio/client";
 import { QUEUE_NAME } from "./config";
 import {
   agentLoopWorkflow,
   compactionWorkflow,
   runSandboxChildToolWorkflow,
-  runSandboxFunctionInvocationWorkflow,
-  runSandboxFunctionToolWorkflow,
 } from "./workflows";
 
 export async function launchAgentLoopWorkflow({
@@ -261,90 +253,6 @@ export async function launchSandboxChildToolWorkflow(
       return new Ok(undefined);
     }
     throw error;
-  }
-
-  return new Ok(undefined);
-}
-
-export async function launchSandboxFunctionToolWorkflow(
-  auth: Authenticator,
-  { action }: { action: SandboxFunctionMCPActionResource }
-): Promise<Result<undefined, Error>> {
-  const authType = auth.toJSON();
-  const { workspaceId } = authType;
-  const client = await getTemporalClientForAgentNamespace();
-
-  const workflowId = makeSandboxFunctionToolWorkflowId({
-    workspaceId,
-    actionModelId: action.id,
-  });
-
-  try {
-    await client.workflow.start(runSandboxFunctionToolWorkflow, {
-      args: [{ authType, actionModelId: action.id }],
-      taskQueue: QUEUE_NAME,
-      workflowId,
-      searchAttributes: {
-        workspaceId: [workspaceId],
-      },
-      memo: {
-        workspaceId,
-      },
-    });
-  } catch (error) {
-    if (error instanceof WorkflowExecutionAlreadyStartedError) {
-      // Idempotent: another caller already kicked it off for this action.
-      return new Ok(undefined);
-    }
-    throw error;
-  }
-
-  return new Ok(undefined);
-}
-
-export async function launchSandboxFunctionInvocationWorkflow(
-  auth: Authenticator,
-  {
-    sandboxFunction,
-    invocation,
-  }: {
-    sandboxFunction: SandboxFunctionResource;
-    invocation: SandboxFunctionInvocationResource;
-  }
-): Promise<Result<undefined, Error>> {
-  const authType = auth.toJSON();
-  const { workspaceId } = authType;
-  const workflowId = makeSandboxFunctionInvocationWorkflowId({
-    workspaceId,
-    invocationId: invocation.sId,
-  });
-
-  try {
-    const client = await getTemporalClientForAgentNamespace();
-    await client.workflow.start(runSandboxFunctionInvocationWorkflow, {
-      args: [
-        {
-          authType,
-          sandboxFunctionId: sandboxFunction.sId,
-          invocationId: invocation.sId,
-        },
-      ],
-      taskQueue: QUEUE_NAME,
-      workflowId,
-      searchAttributes: {
-        workspaceId: [workspaceId],
-      },
-      memo: {
-        sandboxFunctionId: sandboxFunction.sId,
-        invocationId: invocation.sId,
-        workspaceId,
-      },
-    });
-  } catch (error) {
-    if (error instanceof WorkflowExecutionAlreadyStartedError) {
-      return new Ok(undefined);
-    }
-    return new Err(normalizeError(error));
   }
 
   return new Ok(undefined);
