+++ front/temporal/agent_loop/lib/workflow_ids.ts
@@ -42,23 +42,3 @@ export function makeSandboxChildToolWorkflowId({
 }) {
   return `sandbox-child-tool-workflow-${workspaceId}-${actionModelId}`;
 }
-
-export function makeSandboxFunctionToolWorkflowId({
-  workspaceId,
-  actionModelId,
-}: {
-  workspaceId: string;
-  actionModelId: number;
-}) {
-  return `sandbox-function-tool-workflow-${workspaceId}-${actionModelId}`;
-}
-
-export function makeSandboxFunctionInvocationWorkflowId({
-  workspaceId,
-  invocationId,
-}: {
-  workspaceId: string;
-  invocationId: string;
-}) {
-  return `sandbox-function-invocation-workflow-${workspaceId}-${invocationId}`;
-}
