+++ front/temporal/sandbox_functions/config.ts
@@ -1,0 +1,3 @@
+const QUEUE_VERSION = 1;
+
+export const QUEUE_NAME = `sandbox-functions-queue-v${QUEUE_VERSION}`;
